requirements.txt
alembic==1.7.5
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1